requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic>=2.0
pydantic-settings>=2.0
numpy>=2.0
torch>=2.0
transformers>=5.0
faiss-cpu>=1.13
scipy>=1.17
scikit-learn>=1.8
